Add lint for Golang test size tags
