+++ api/manifest_test.go
@@ -1,3 +1,5 @@
+// +build small
+
 // Copyright 2017 The WPT Dashboard Project. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
@@ -5,35 +7,12 @@
 package api
 
 import (
-	"bytes"
-	"compress/gzip"
 	"encoding/json"
 	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
-
-const fullSHA = "abcdef0123456789abcdef0123456789abcdef01"
-
-// Shorthand for arbitrary json objects.
-type object map[string]interface{}
-
-type mockGitHubClient struct {
-	Responses map[string][]byte
-}
-
-func (m *mockGitHubClient) fetch(url string) ([]byte, error) {
-	if _, ok := m.Responses[url]; !ok {
-		return nil, fmt.Errorf("fore! oh; for: %s", url)
-	}
-	return m.Responses[url], nil
-}
-
-func unsafeMarshal(i interface{}) []byte {
-	result, _ := json.Marshal(i)
-	return result
-}
 
 func TestGetGitHubReleaseAssetForSHA_SHANotFound(t *testing.T) {
 	client := mockGitHubClient{}
@@ -126,12 +105,4 @@ func TestGetGitHubReleaseAssetLatest(t *testing.T) {
 	sha, manifestGZIP, _ = getGitHubReleaseAssetForSHA(&client, "latest")
 	assert.Equal(t, data, manifestGZIP)
 	assert.Equal(t, fullSHA, sha)
-}
-
-func getManifestPayload(data string) []byte {
-	var buf bytes.Buffer
-	zw := gzip.NewWriter(&buf)
-	zw.Write([]byte(data))
-	zw.Close()
-	return buf.Bytes()
 }
