+++ api/results_redirect_handler_test.go
@@ -1,3 +1,5 @@
+// +build small
+
 // Copyright 2017 The WPT Dashboard Project. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
