+++ shared/tag_test.go
@@ -1,0 +1,7 @@
+// +build !small,!medium,!large
+
+package shared
+
+func init() {
+	panic("Tests were run without -tags=[small|medium|large]")
+}
